requirements.txt
anyascii==0.1.7
asgiref>=3.8.0
beautifulsoup4==4.8.2
certifi==2020.12.5
chardet==4.0.0
convertdate==2.4.0
Django>=4.2,<5.0
django-admin-sortable>=2.3
django-anymail==8.1
django-cors-headers==3.6.0
django-extensions==3.1.0
django-filter>=23.3.0
django-modelcluster>=6.0
django-recaptcha>=2.0.6
django-social-share==2.1.0
django-taggit>=5.0,<5.2
django-timezone-field==5.0
django-treebeard>=4.5.1
django-widget-tweaks==1.4.8
djangorestframework>=3.15
docopt==0.6.2
draftjs-exporter==2.1.7
et-xmlfile==1.0.1
feedgen==0.9.0
hijri-converter==2.2.3
holidays==0.12
html5lib==1.1
l18n==2020.6.1
lxml==4.8.0
num2words==0.5.10
openpyxl>=3.0.10
Pillow>=9.1
psycopg2>=2.9
PyMeeus==0.5.11
python-dateutil==2.8.2
pytz>=2022.4
requests==2.25.1
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
tablib==3.0.0
telepath>=0.3.1
Unidecode==1.1.1
urllib3==1.26.2
wagtail>=6.0
wagtail-gallery==0.1.1
wagtailgmaps==1.0.1
webencodings==0.5.1
Willow>=1.8
xlrd==2.0.1
XlsxWriter==1.3.7
xlwt==1.3.0
